+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADGroup/MSFT_AADGroup.psm1
@@ -51,6 +51,10 @@ function Get-TargetResource
         $Visibility,
 
         [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance[]]
+        $AssignedLicenses,
+
+        [Parameter()]
         [ValidateSet('Present', 'Absent')]
         [System.String]
         $Ensure = 'Present',
@@ -133,6 +137,14 @@ function Get-TargetResource
         {
             Write-Verbose -Message "Found existing AzureAD Group"
 
+            $assignedLicensesRequest = Invoke-MgGraphRequest -Method 'GET' `
+                -Uri "https://graph.microsoft.com/v1.0/groups/$($Group.Id)/assignedLicenses"
+
+            if ($assignedLicensesRequest.value.Length -gt 0)
+            {
+                $assignedLicensesValues = Get-M365DSCAzureADGroupLicenses -AssignedLicenses $assignedLicensesRequest.value
+            }
+
             $result = @{
                 DisplayName                   = $Group.DisplayName
                 Id                            = $Group.Id
@@ -145,6 +157,7 @@ function Get-TargetResource
                 IsAssignableToRole            = $Group.IsAssignableToRole
                 MailNickname                  = $Group.MailNickname
                 Visibility                    = $Group.Visibility
+                AssignedLicenses              = $assignedLicensesValues
                 Ensure                        = "Present"
                 ApplicationId                 = $ApplicationId
                 TenantId                      = $TenantId
@@ -230,6 +243,10 @@ function Set-TargetResource
         $Visibility,
 
         [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance[]]
+        $AssignedLicenses,
+
+        [Parameter()]
         [ValidateSet('Present', 'Absent')]
         [System.String]
         $Ensure = 'Present',
@@ -285,10 +302,93 @@ function Set-TargetResource
         Write-Verbose -Message "Cannot set mailenabled to false if GroupTypes is set to Unified when creating group."
         throw "Cannot set mailenabled to false if GroupTypes is set to Unified when creating a group."
     }
-    if (-not $GroupTypes)
+    if (-not $GroupTypes -and $currentParameters.GroupTypes -eq $null)
     {
         $currentParameters.Add("GroupTypes", @("Unified"))
     }
+
+    $currentValuesToCheck = @()
+    if ($currentGroup.AssignedLicenses.Length -gt 0)
+    {
+        $currentValuesToCheck = $currentGroup.AssignedLicenses.SkuId
+    }
+    $desiredValuesToCheck = @()
+    if ($AssignedLicenses.Length -gt 0)
+    {
+        $desiredValuesToCheck = $AssignedLicenses.SkuId
+    }
+
+    [Array]$licensesDiff = Compare-Object -ReferenceObject $currentValuesToCheck -DifferenceObject $desiredValuesToCheck -IncludeEqual
+    $toAdd = @()
+    $toRemove = @()
+    foreach ($diff in $licensesDiff)
+    {
+        if ($diff.SideIndicator -eq '=>')
+        {
+            $toAdd += $diff.InputObject
+        }
+        elseif ($diff.SideIndicator -eq '<=')
+        {
+            $toRemove += $diff.InputObject
+        }
+        elseif ($diff.SideIndicator -eq '==')
+        {
+            # This will take care of the scenario where the license is already assigned but has different disabled plans
+            $toAdd += $diff.InputObject
+        }
+    }
+
+    # Convert AssignedLicenses from SkuPartNumber back to GUID
+    $licensesToAdd = @()
+    $licensesToRemove = @()
+    [Array]$AllLicenses = Get-M365DSCCombinedLicenses -DesiredLicenses $AssignedLicenses -CurrentLicenses $currentGroup.AssignedLicenses
+
+    $allSkus = Get-MgSubscribedSku
+    # Create complete list of all Service Plans
+    $allServicePlans = @()
+    Write-Verbose -Message "Getting all Service Plans"
+    foreach ($sku in $allSkus)
+    {
+        foreach ($serviceplan in $sku.ServicePlans)
+        {
+            if ($allServicePlans.Length -eq 0 -or -not $allServicePlans.ServicePlanName.Contains($servicePlan.ServicePlanName))
+            {
+                $allServicePlans += @{
+                    ServicePlanId   = $serviceplan.ServicePlanId
+                    ServicePlanName = $serviceplan.ServicePlanName
+                }
+            }
+        }
+    }
+
+    foreach ($assignedLicense in $AllLicenses)
+    {
+        $skuInfo = $allSkus | Where-Object -FilterScript {$_.SkuPartNumber -eq $assignedLicense.SkuId}
+        if ($skuInfo)
+        {
+            if ($toAdd.Contains($assignedLicense.SkuId))
+            {
+                $disabledPlansValues = @()
+                foreach ($plan in $assignedLicense.DisabledPlans)
+                {
+                    $foundItem = $allServicePlans | Where-Object -FilterScript {$_.ServicePlanName -eq $plan}
+                    $disabledPlansValues += $foundItem.ServicePlanId
+                }
+
+                $skuInfo = $allSkus | Where-Object -FilterScript {$_.SkuPartNumber -eq $assignedLicense.SkuId}
+                $licensesToAdd += @{
+                    DisabledPlans = $disabledPlansValues
+                    SkuId         = $skuInfo.SkuId
+                }
+            }
+            elseif ($toRemove.Contains($assignedLicense.SkuId))
+            {
+                $licensesToRemove += $skuInfo.SkuId
+            }
+        }
+    }
+
+    $currentParameters.Remove('AssignedLicenses') | Out-Null
 
     if ($Ensure -eq 'Present' -and $currentGroup.Ensure -eq 'Present')
     {
@@ -307,12 +407,30 @@ function Set-TargetResource
             }
             else
             {
-                Write-Verbose -Message "Updating settings for group {$DisplayName}"
+                $currentParameters.Remove("Id") | Out-Null
+                $currentParameters.Add("GroupId", $currentGroup.Id)
+                Write-Verbose -Message "Updating Group with Values: $(Convert-M365DscHashtableToString -Hashtable $currentParameters)"
                 Update-MgGroup @currentParameters | Out-Null
+            }
+
+            if ($licensesToAdd.Length -gt 0 -or $licensesToRemove.Length -gt 0)
+            {
+                try
+                {
+                    Set-MgGroupLicense -GroupId $currentGroup.Id `
+                        -AddLicenses $licensesToAdd `
+                        -RemoveLicenses $licensesToRemove `
+                        -ErrorAction Stop | Out-Null
+                }
+                catch
+                {
+                    Write-Verbose -Message $_
+                }
             }
         }
         catch
         {
+            Write-Error -Message $_
             New-M365DSCLogEntry -Error $_ -Message "Couldn't set group $DisplayName" -Source $MyInvocation.MyCommand.ModuleName
         }
     }
@@ -320,9 +438,17 @@ function Set-TargetResource
     {
         Write-Verbose -Message "Creating new group {$DisplayName}"
         $currentParameters.Remove("Id") | Out-Null
+
         try
         {
-            New-MgGroup @currentParameters | Out-Null
+            Write-Verbose -Message "Creating Group with Values: $(Convert-M365DscHashtableToString -Hashtable $currentParameters)"
+            $group = New-MgGroup @currentParameters
+
+            Write-Verbose -Message "Created Group $($group.id)"
+            if ($assignedLicensesGUIDs.Length -gt 0)
+            {
+                Set-MgGroupLicense -GroupId $group.Id -AddLicenses $licensesToAdd -RemoveLicenses @()
+            }
         }
         catch
         {
@@ -396,6 +522,10 @@ function Test-TargetResource
         $Visibility,
 
         [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance[]]
+        $AssignedLicenses,
+
+        [Parameter()]
         [ValidateSet('Present', 'Absent')]
         [System.String]
         $Ensure = 'Present',
@@ -439,12 +569,60 @@ function Test-TargetResource
 
     Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $PSBoundParameters)"
 
+    # Check Licenses
+    try
+    {
+        $licensesDiff = Compare-Object -ReferenceObject ($CurrentValues.AssignedLicenses.SkuId) -DifferenceObject ($AssignedLicenses.SkuId)
+        if ($null -ne $licensesDiff)
+        {
+            Write-Verbose -Message "AssignedLicenses differ: $($licensesDiff | Out-String)"
+            Write-Verbose -Message "Test-TargetResource returned $false"
+            $EventMessage = "Assigned Licenses for Azure AD Group {$DisplayName} were not in the desired state.`r`n" + `
+                "They should contain {$($AssignedLicenses.SkuId)} but instead contained {$($CurrentValues.AssignedLicenses.SkuId)}"
+            Add-M365DSCEvent -Message $EventMessage -EntryType 'Warning' `
+                -EventID 1 -Source $($MyInvocation.MyCommand.Source)
+            return $false
+        }
+        else
+        {
+            Write-Verbose -Message "AssignedLicenses for Azure AD Group are the same"
+        }
+    }
+    catch
+    {
+        Write-Verbose -Message "Test-TargetResource returned $false"
+        return $false
+    }
+
+    #Check DisabledPlans
+    try
+    {
+        $licensesDiff = Compare-Object -ReferenceObject ($CurrentValues.AssignedLicenses.DisabledPlans) -DifferenceObject ($AssignedLicenses.DisabledPlans)
+        if ($null -ne $licensesDiff)
+        {
+            Write-Verbose -Message "DisabledPlans differ: $($licensesDiff | Out-String)"
+            Write-Verbose -Message "Test-TargetResource returned $false"
+            $EventMessage = "Disabled Plans for Azure AD Group Licenses {$DisplayName} were not in the desired state.`r`n" + `
+                "They should contain {$($AssignedLicenses.DisabledPlans)} but instead contained {$($CurrentValues.AssignedLicenses.DisabledPlans)}"
+            Add-M365DSCEvent -Message $EventMessage -EntryType 'Warning' `
+                -EventID 1 -Source $($MyInvocation.MyCommand.Source)
+            return $false
+        }
+        else
+        {
+            Write-Verbose -Message "DisabledPlans for Azure AD Group Licensing are the same"
+        }
+    }
+    catch
+    {
+        Write-Verbose -Message "Test-TargetResource returned $false"
+        return $false
+    }
+
     $ValuesToCheck = $PSBoundParameters
-    $ValuesToCheck.Remove('ApplicationId') | Out-Null
-    $ValuesToCheck.Remove('TenantId') | Out-Null
-    $ValuesToCheck.Remove('ApplicationSecret') | Out-Null
     $ValuesToCheck.Remove('Id') | Out-Null
     $ValuesToCheck.Remove('GroupTypes') | Out-Null
+    $ValuesToCheck.Remove('AssignedLicenses') | Out-Null
 
     $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
         -Source $($MyInvocation.MyCommand.Source) `
@@ -519,11 +697,20 @@ function Export-TargetResource
             $Results = Get-TargetResource @Params
             $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `
                 -Results $Results
+            if ($results.AssignedLicenses.Length -gt 0)
+            {
+                $Results.AssignedLicenses = Get-M365DSCAzureADGroupLicensesAsString $Results.AssignedLicenses
+            }
             $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
                 -ConnectionMode $ConnectionMode `
                 -ModulePath $PSScriptRoot `
                 -Results $Results `
                 -Credential $Credential
+            if ($null -ne $Results.AssignedLicenses)
+            {
+                $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock `
+                    -ParameterName "AssignedLicenses"
+            }
             $dscContent += $currentDSCBlock
             Save-M365DSCPartialExport -Content $currentDSCBlock `
                 -FileName $Global:PartialExportFileName
@@ -556,4 +743,136 @@ function Export-TargetResource
     }
 }
 
+function Get-M365DSCAzureADGroupLicenses
+{
+    [CmdletBinding()]
+    [OutputType([PSCustomObject])]
+    param(
+        [Parameter(Mandatory = $true)]
+        $AssignedLicenses
+    )
+
+    $returnValue = @()
+    $allSkus = Get-MgSubscribedSku
+
+    # Create complete list of all Service Plans
+    $allServicePlans = @()
+    Write-Verbose -Message "Getting all Service Plans"
+    foreach ($sku in $allSkus)
+    {
+        foreach ($serviceplan in $sku.ServicePlans)
+        {
+            if ($allServicePlans.Length -eq 0 -or -not $allServicePlans.ServicePlanName.Contains($servicePlan.ServicePlanName))
+            {
+                Write-Verbose -Message "$($servicePlan.ServicePlanId) :: $($serviceplan.ServicePlanName)"
+                $allServicePlans += @{
+                    ServicePlanId   = $serviceplan.ServicePlanId
+                    ServicePlanName = $serviceplan.ServicePlanName
+                }
+            }
+        }
+    }
+
+    foreach ($assignedLicense in $AssignedLicenses)
+    {
+        $skuPartNumber = $allSkus | Where-Object -FilterScript {$_.SkuId -eq $assignedLicense.SkuId}
+        $disabledPlansValues = @()
+        foreach ($plan in $assignedLicense.DisabledPlans)
+        {
+            $foundItem = $allServicePlans | Where-Object -FilterScript {$_.ServicePlanId -eq $plan}
+            $disabledPlansValues += $foundItem.ServicePlanName
+        }
+        $currentLicense = @{
+            DisabledPlans = $disabledPlansValues
+            SkuId         = $skuPartNumber.SkuPartNumber
+        }
+        $returnValue += $currentLicense
+    }
+
+    return $returnValue
+}
+
+function Get-M365DSCAzureADGroupLicensesAsString
+{
+    [CmdletBinding()]
+    [OutputType([System.String])]
+    param(
+        [Parameter(Mandatory = $true)]
+        [System.Collections.ArrayList]
+        $AssignedLicenses
+    )
+
+    $StringContent = [System.Text.StringBuilder]::new()
+    $StringContent.Append("@(") | Out-Null
+    foreach ($assignedLicense in $AssignedLicenses)
+    {
+        $StringContent.Append("MSFT_AADGroupLicense { `r`n") | Out-Null
+        if ($assignedLicense.DisabledPlans.Length -gt 0)
+        {
+            $StringContent.Append("                DisabledPlans = @('" + ($assignedLicense.DisabledPlans -join "','") + "')`r`n") | Out-Null
+        }
+        else
+        {
+            $StringContent.Append("                DisabledPlans = @()`r`n") | Out-Null
+        }
+        $StringContent.Append("                SkuId         = '" + $assignedLicense.SkuId + "'`r`n") | Out-Null
+        $StringContent.Append("            }`r`n") | Out-Null
+    }
+    $StringContent.Append("            )") | Out-Null
+    return $StringContent.ToString()
+}
+
+function Get-M365DSCCombinedLicenses
+{
+    [CmdletBinding()]
+    [OutputType([System.Object[]])]
+    param(
+        [Parameter()]
+        [System.Object[]]
+        $CurrentLicenses,
+
+        [Parameter()]
+        [System.Object[]]
+        $DesiredLicenses
+    )
+    $result = @()
+    if ($currentLicenses.Length -gt 0)
+    {
+        foreach ($license in $CurrentLicenses)
+        {
+            Write-Verbose -Message "Including Current $license"
+            $result += @{
+                SkuId         = $license.SkuId
+                DisabledPlans = $license.DisabledPlans
+            }
+        }
+    }
+
+    if ($DesiredLicenses.Length -gt 0)
+    {
+        foreach ($license in $DesiredLicenses)
+        {
+            if (-not $result.SkuId.Contains($license.SkuId))
+            {
+                $result += @{
+                    SkuId         = $license.SkuId
+                    DisabledPlans = $license.DisabledPlans
+                }
+            }
+            else
+            {
+                #Set the Desired Disabled Plans if the sku is already added to the list
+                foreach ($item in $result)
+                {
+                    if ($item.SkuId -eq $license.SkuId)
+                    {
+                        $item.DisabledPlans = $license.DisabledPlans
+                    }
+                }
+            }
+        }
+    }
+
+    return $result
+}
 Export-ModuleMember -Function *-TargetResource